% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{broward}
\alias{broward}
\title{Disaster declarations in Broward County, Florida}
\format{
A data frame with 20 rows and 6 variables:
\describe{
  \item{fipsStateCode}{FIPs code for states}
  \item{fipsCountyCode}{FIPS code for counties}
  \item{designatedArea}{Name of Area}
  \item{fyDeclared}{Year of disaster declaration}
  \item{incidentType}{Type of disasater declaration}
  \item{declarations}{Number of declarations, type per year}
}
}
\source{
\url{}
}
\usage{
broward
}
\description{
A dataset containing the number of recruits per 100,000 people compared to the percentage of people of color in New York City neighborhoods.
Originally from FEMA's open data files but was cleaned for use in course.
Example using this data was also used in her book.
}
\keyword{datasets}
