% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quartets}
\alias{quartets}
\title{Anacombe's Qaurtets (1973)}
\format{
A data frame with 44 observations and 3 variables
\describe{
  \item{group}{a categorical identifier for the quartet}
  \item{x}{a continuous variable}
  \item{y}{a continuous variable}
}
}
\source{
\url{}
}
\usage{
quartets
}
\description{
Contains 4 x-y datasets in long format.
Have same mean,variance, correlation, regression line, etc.
But look very different.
}
\keyword{datasets}
