% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recruits}
\alias{recruits}
\title{Prices of 50,000 round cut diamonds.}
\format{
A data frame with 190 rows and 4 variables:
\describe{
  \item{Borough}{Borough acronym}
  \item{Area Name/Zip Code}{}
  \item{Total Recruits per 100,000}{Number of individuals recruited per 100,000 residents}
  \item{Percent Persons of Color}{Percentage of population that is a person of color}
}
}
\source{
\url{https://stephanieevergreen.com/books/}
}
\usage{
recruits
}
\description{
A dataset containing the number of recruits per 100,000 people compared to the percentage of people of color in New York City neighborhoods.
Originally from http://www.radicalmath.org/ but that site is no longer maintained.
Data are from Stephanie Evergreen's practice file for her book "Effective Data Visualization"
Example using this data was also used in her book.
}
\keyword{datasets}
